Working Text Fields for Data Input
Can input text in data field entry

Also text inputs can have string limits, this is to prevent any extending of boxes

Has separate ones for 6x6, 9x9, 12x12

Working on rigging to Generate Button, ran into an interesting scenario of print statement looping repeatedly. Want to see if it does so in Main as well so I know if it's something I messed up or not.